rcdom: Port remaining tests to a custom runner
Move the custom test runner into its own `util::runner` module
and use it instead of rustc_test for the other test files.
This allows `cargo test` to complete under rust 1.77.1.
