+++ rcdom/Cargo.toml
@@ -22,7 +22,6 @@ xml5ever = { version = "0.18", path = "../xml5ever" }
 
 [dev-dependencies]
 serde_json = "1.0"
-rustc-test = "0.3"
 
 [[test]]
 name = "html-tokenizer"
